requirements.txt
xformers==0.0.26.post1+cu118
torchvision
absl-py
tensorboard
lpips
kornia
opencv-python
ffmpeg
einops
beartype
imageio[ffmpeg]
blobfile
ninja
pyspng
psutil
mrcfile
lmdb
webdataset
point_cloud_utils
kiui
plyfile
open_clip_torch==2.24.0
openai-clip
timm
matplotlib
torchtyping
omegaconf
pytorch_lightning
lz4
xatlas
open3d
trimesh
huggingface_hub
safetensors
PyMCubes
vision aided loss
transformers
torchdiffeg==8.2.4torchmetrics==1.4.3torchtyping==0.1.5imath
OpenEXR
onnxscript
tritonclient
third_party/diff-surfel-rasterization